backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
groq
python-dotenv
python-multipart
